test(runtime): split config policy coverage
